+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support for loading components that are ES Modules. Note that they must directly import `noflo/src/lib/Component` instead of `noflo` to prevent circular dependencies:
+
+```javascript
+import { Component } from 'noflo/src/lib/Component.js';
+
+export function getComponent() {
+  const c = new Component();
+  // Implementation
+  return c;
+}
+```
+
 ## [1.5.1] - 2026-06-20
 ### Changed
 - NoFlo can now be imported either ES Module or CommonJS